# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Scrap Payments Manager API",
        "description": "API documentation for Scrap Payments Manager application",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000api/v1",
            "description": "API Server"
        }
    ],
    "paths": {
...
        "/dashboard/stats": {
            "get": {
                "tags": [
                    "Dashboard"
                ],
                "summary": "Obtener estadísticas generales del dashboard",
                "operationId": "06b55ec316b401263ae54a499fc977ad",
                "responses": {
                    "200": {
                        "description": "Operación exitosa"
                    },
                    "401": {
                        "description": "No autenticado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/dashboard/recent-activity": {
            "get": {
                "tags": [
                    "Dashboard"
                ],
                "summary": "Obtener actividad reciente (créditos y pagos)",
                "operationId": "dcbdaa7960b57a1404fb0adddc9e7302",
                "parameters": [
                    {
                        "name": "limit",
                        "in": "query",
                        "description": "Número de registros a retornar",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 10
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Operación exitosa"
                    },
                    "401": {
                        "description": "No autenticado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/dashboard/monthly-overview": {
            "get": {
                "tags": [
                    "Dashboard"
                ],
                "summary": "Obtener datos para el gráfico de resumen mensual",
                "operationId": "595842cfb0738db2465642487fa61ec4",
                "parameters": [
                    {
                        "name": "months",
                        "in": "query",
                        "description": "Número de meses a retornar",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 6
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Operación exitosa"
                    },
                    "401": {
                        "description": "No autenticado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/dashboard/scrap-collection": {
            "get": {
                "tags": [
                    "Dashboard"
                ],
                "summary": "Estadísticas de recolección de chatarra por periodo",
                "operationId": "a7135fe3e9954ed39f1b61c5b3078253",
                "parameters": [
                    {
                        "name": "startDate",
                        "in": "query",
                        "description": "Fecha inicio (YYYY-MM-DD)",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "date"
                        }
                    },
                    {
                        "name": "endDate",
                        "in": "query",
                        "description": "Fecha fin (YYYY-MM-DD)",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "date"
                        }
                    },
                    {
                        "name": "scrapId",
                        "in": "query",
                        "description": "ID de chatarra opcional",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Operación exitosa"
                    },
                    "401": {
                        "description": "No autenticado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Enter token in format (Bearer <token>)",
                "scheme": "bearer",
                "bearerFormat": "JWT"
            }
        }
    },
...
}